docker_compose/simple_project/app/requirements.txt
django==4.0.4
uwsgi==2.0.20
python-dotenv
split_settings
psycopg2-binary